Delete TabMenuModelDelegateProxy
Instead, BraveTabMenuModel acts as a proxy of TabContextMenuContents.
